Make it easier to build for arm
Add a nacl build tag to use the golang.org/x/crypto/nacl library instead
of github.com/jamesruan/sodium. The latter uses CGO, which is a pain to
cross-compile. Fortunately, the x/crypto/nacl library seems to work just
as well.

Now, it's possible to cross-compile for ARM (or other platforms) easily.