Compat: keep filenames the same size
The Stingle App assumes that headers don't change size, which makes
renaming files a little bit tricky.

To make renames work and keep the app happy, we'll use a fixed size for
filenames and pad them with leading spaces. This is almost transparent.
It's only visible when the app decrypts a file and the padded filename
becomes visible in android.