Remove support for old server db formats.
Staring with this commit, it is expected that all databases have been
converted to the latest format with:

  inspect encrypt-blobs
  inspect rename-ref-files
  inspect flatten-data-dir

Note that these commands are removed in this commit. So, inspect would
need to be built at the previous commit to include these subcommands,
i.e.
  $ git checkout 559b7ca7ee6eb368e190f0827d7b6369ca0f7d43
  $ export C2FMZQ_DATABASE="<database dir>"
  $ go run ./c2FmZQ-server/inspect encrypt-blobs
  $ go run ./c2FmZQ-server/inspect rename-ref-files
  $ go run ./c2FmZQ-server/inspect flatten-data-dir

Make sure you have a backup of your database directory, just in case.