Experiment with duress/decoy passwords.
Decoy passwords can be associated with any normal account. When the decoy
password is used to login, the login is successful, but the user is
actually logged in with a different account.

Note that decoy accounts are not as safe as normal accounts because the
passwords have to be known by the server. So, someone with access to the
server metadata could access the files in any decoy account.